requirements.txt
asgiref==3.2.10
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
django-cors-headers==3.4.0
django-filter==2.3.0
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
psycopg2==2.8.5
django-dbbackup==3.3.0
whitenoise==5.0.1
# Markdown==3.1.1
# numpy==1.19.0
# pandas==0.25.2
# requests==2.22.0
# scikit-learn==0.21.3
# # Cython
# TensorFlow==2.9.0